node-packages: Regenerate with node2nix 1.6.0
Well, it's not exactly *fully* regenerated, but only so that we have the
fixes to node-env.nix. The package versions are still the same, except
nodemailer being removed (from the mailer-daemon.patch).

The reason why I didn't do a full update is that upstream didn't do it
since a while, so I'm suspecting that there is a reason for that and I
don't want break anything right now.

Signed-off-by: aszlig <aszlig@nix.build>