tests/e2e: Pin Selenium/Chromium
I've tested the e2e test only on NixOS Unstable and after watching the
Hydra tests, they seem to be failing with the Selenium/Chromium variant
we have in NixOS 18.03.

So in order to make the test stable, I have pinned the nixpkgs version
to one that contains the Selenium/Chromium versions that work with the
tests. We don't have to care about security updates or having old
versions in there, because it's only for the E2E test anyway and not
used elsewhere.

Signed-off-by: aszlig <aszlig@nix.build>