imageproxy: Test with real DNS resolver
So far our test used /etc/hosts based lookup, which isn't really what we
want to have in the real world. Therefore I have removed /etc/hosts and
a bunch of other files in /etc that we don't need from the chroot.

The test now uses a real DNS server which ships a fake root zone with
all the addresses of the nodes.

Signed-off-by: aszlig <aszlig@nix.build>