imageproxy: Don't cache errors at all
The test case for this was actually wrong and I've fixed this so that
the URL passed to the proxy now is correct but the session isn't.

While fixing the test case, I found that the HTTP 403 that was returned
was cached by nginx, so let's actually avoid that in our nginx config.

Signed-off-by: aszlig <aszlig@nix.build>