Update to version 4.56.3
Upstream API changes:

  * Fix: Localize timestamps for due dates and chat messages
  * Fix: Prevent duplicate group chat notifications from being added to
         user's list
  * Fix: Don't let newly created accounts hide chat messages via
         flagging, to prevent trollish flag exploitation
  * Chore: Locale files updated

Upstream Client changes:

  * Fix: Correctly link to Tavern or Party when creating To-Dos from
         chat messages, as appropriate
  * Fix: Localize text shown in notifications panel when user has no
         remaining notifications

Other upstream changes:

  * Fix: Set development mode correctly in docker-compose

The changes for 4.56.1 and 4.56.2 are only news announcements and
translation updates, so not really relevant to us.

While rebasing the patches I had to remove a bunch of code in
POST-chat.flag.test.js because they contain references to the Slack
client library.

Signed-off-by: aszlig <aszlig@nix.build>