Update to version 4.57.0
Upstream API changes:

  * Content: Kangaroo Pet Quest
  * Docs: Improvements regarding Stat Points, user, and unlock
  * Chore: Locale files updated
  * Chore: Bailey news announcements

Upstream client changes:

  * Fix: Display Experience gain accurately when user levels up

Upstream repository changes:

  * Chore: Update various packages as advised by Greenkeeper

The last point about the update of the various packages is not really
true, because it has been reverted thereafter. However one packaging
change is there, which is the downgrade of css-loader to 0.28.11.

We did not update to version 1.0.0, so we don't need to downgrade on our
side.

Signed-off-by: aszlig <aszlig@nix.build>