Rename almost everything to Shabitica
I have just announced[1] this on the NixOS Discourse, so before users
start to actually use this, let's rename the NixOS options and all the
attributes accordingly, as well as the Hydra jobsets (we have two now,
one for NixOS 18.03 and one for NixOS Unstable).

This was long overdue and the repository has this name already. It
probably would have been wiser to do this *prior* to the announcement,
so I could do this without hurry. But well...

I only evaluated the tests but didn't build them, if they fail to build
it's not a big deal as we can fix this within the next few minutes and
it's more important to have the now attribute names in place.

[1]: https://discourse.nixos.org/t/self-hosted-habitica/735

Signed-off-by: aszlig <aszlig@nix.build>