source: Add missing function canaries
I didn't run find-canaries.py for a while after rebasing patches, so two
canaries were missing (actually three, one for the "notify" function as
well, but that one is to generic and causes false positives).

Apart from just adding the canaries, I also updated the comment above
the version so I don't forget to do it next time. In the long run it
might be a good idea to automate this further (like eg. during patch
updating).

Signed-off-by: aszlig <aszlig@nix.build>