Update to version 4.59.2
The commit message for the last update actually had the wrong version in
the subject line, so we're really updating from 4.59.0 to 4.59.2.

Upstream API changes:

  * Fix: Don't push 0-damage notifications into party chat when questing
         party member is in the Inn
  * Fix: Make Lost Masterclasser completionist achievement more reliable
  * Fix: Further improvements to reliability of Group Plan member counts
  * Docs: Typo fixes
  * Chore: Locale files updated
  * Chore: Bailey news announcement

Upstream client changes:

  * Feature: Show smart banner when user is on mobile browser instead of
             ugly red warning
  * Fix: Allow social login buttons to expand to fit long text, e.g. for
         translations that run longer than en
  * Fix: Show correct/up-to-date information in members modal for
         Challenges
  * Fix: Continue displaying completed To-Dos after manual sync

These are the changes only for 4.59.1, because 4.59.2 consists of only
reverts, which are:

  * Revert: Show accurate XP gain in notification on level up
  * Revert: Check user version before adding notifications

Apart from their respective commit message subjects, it's not really
clear why they've got reverted but I'm guessing that these changes did
introduce some problems.

When rebasing the patches, I had to remove all the smartbanner code as
we don't package the mobile apps. I also added a canary for
"smartbanner" to make sure that new pieces of it aren't creeping in in
future updates.

Signed-off-by: aszlig <aszlig@nix.build>