Remove "apn" from package.json again
This was introduced again in cd37741e17939bd0172f49d3b26c5ea0aabfa8fe.

The reason I didn't notice this is that the canary check had an
exclusion on package.json, so I removed the exclusion.

Luckily however we are completely ignoring package.json and only use the
node packages generated via node2nix, so that dependency didn't land.

Signed-off-by: aszlig <aszlig@nix.build>