source: Be more selective about canary exclusions
We had a lot of paths that were excluded from the canary search so far,
which has the disadvantage that things *might* slip in there on new
upstream releases.

Now we are trying to keep the exclusions as minimal as possible so that
things like the accidental add of "apn" in package.json (see previous
commit) doesn't happen again.

In addition to that, we now kill off a few more files and patch out more
unused code. For example I initially was quite reluctant when it comes
to removing stuff from the schema, but for MongoDB this really isn't an
issue anyway as those things are only added whenever they're really
used. One example for that would be pushDevices, which we don't use as
we don't support mobile clients.

Related, I also removed the Slack integration in gulp/taskHelper, which
really doesn't have an impact on our end but it's still ugly to have it
laying around in the patched source tree.

Signed-off-by: aszlig <aszlig@nix.build>