Update to version 4.61.0
Relevant upstream updates:

  * Fall Festival 2018
  * Fix pinning equipment pieces

This update also bumps a few dependencies, so I've regenerated them on
our side. However, I needed to get rid of a few overrides because they
are no longer needed and even break the builds.

Also, instead of just using the bcrypt version that is stated in the
package.json of Habitica, we need to change it to use the upstream
version instead, because the version in package.json doesn't have
node-pre-gyp bundled.

Signed-off-by: aszlig <aszlig@nix.build>