node-overrides/apidoc: Prefetch Google fonts
While we were quite thorough when it comes to external scripts and fonts
in Habitica itself, the apidoc went under our radar, because the loading
of the Google fonts is done by the "apidoc" NPM package rather than some
part of the Habitica source code itself.

This was also the reason why no canary was triggered.

So we're now doing something similar to what we do in source.nix, mainly
using "google-fonts-offline" to fetch the CSS and the font files and
patching them into the template.

Signed-off-by: aszlig <aszlig@nix.build>