Fix channel to play nicely with "nix-env -qa"
As reported in #2:

  ... if I add the shabitica channel ... and do a nix-channel --update,
  then calling nix-env -qa as a root user gives:

  error: cannot auto-call a function that has an argument without a
  default value ('config')

We have two options here: Either move the NixOS module to another file
or make sure that whenever the module is called via nix-env it won't
result in an error.

While the former option would in theory be the easiest solution I didn't
choose to do so for two reasons:

  * Having something like <shabitica/module.nix> in the module imports
    is unnecessarily adding another file name to remember and it's also
    superfluous, because all we want to provide via the channel is the
    module. If we want to provide other things, we can still add another
    channel.
  * Changing the path would mean to break existing deployments and while
    the project probably isn't widely deployed I don't have metrics and
    I want to avoid breaking things even if it is for just one user.

The latter option however does play nicely with nix-env and in order to
make sure I accidentally break things I added a comment for now. In the
long term it's probably also a good idea to create a test specifically
for interacting with the channel.

Signed-off-by: aszlig <aszlig@nix.build>
Fixes: #2