Update to version 4.65.2
This update changes quite a lot given that I missed the last few updates
and I was away during the weekend.

The biggest change is the username refactor, which introduced a lot of
conflicts in the patches due to code that has been moved around.

Along these changes, a new dependency "short-uuid" has been introduced
which is used for the generateUsername function. I've removed that
function and the dependency because it is only used for social auth
which we don't use altogether.

Banned words/slurs now also apply to usernames, so I removed that
functionality as well.

Another change is that the private messages now reside in their own
collection and we need another migration. This also moves around the
Chat model into a new source file, so we needed to fix the import of the
last migration to point to the new file.

I also had to fix a few API documentation related errors, which I also
submitted upstream at https://github.com/HabitRPG/habitica/pull/10749.

Now the habitipy package gets passed shabiticaSource attribute, so that
we can use API v4 in the tests of the new migration. The whole reason we
need support for API v4 is because the API v3 version includes the inbox
messages in the user object even though it's in a different collection
and we want to verify this via the API after the migration is done.

Here is the list of relevant changes without locale, news updates and
GitHub mentions:

Version 4.65.0:

  * Content: Alligator Pets and Spoopy Sporples
  * Fix: Group plan username verification

Version 4.64.2:

  * Fix: Add spritesheet 23 to app.vue

Version 4.64.1:

  * Fix: Don't throw "already taken" error if user changes/verifies
         username as case-insensitive variant of existing name

Version 4.64.0:

  * Content: New backgrounds and Enchanted Armoire items for October
             2018
  * Fix: Upgrade Facebook Graph API for social login flow to current
         version

Version v4.63.3:

  * Fix: Address 500 error that could occur when registering a new
         account with local authentication

Version 4.63.2:

  * API fix: Reward error when user enters invalid username characters,
             for clarity
  * API fix: Correct typo in error when user attempts to include a slur
             in a username
  * Client fix: Remove "Confirm" button that was confusing users, and
                merge its functionality with the "Submit" button as a
                "Save & Confirm" button
  * Client fix: If user leaves username field empty, restore its
                previous state on loss of focus
  * Client fix: Account for preexisting usernames in functionality for
                adding local authentication

Version 4.63.0:

  * Feature: New handling for auth.local.username field. All users can
             now employ this field, including those logging in with
             social authentication. Includes a migration for assigning
             an automatically generated username to users, a workflow
             and field for marking one's username as verified, and a
             script for sending an email announcement informing them of
             this change, because this lays the groundwork for later
             features built on usernames--requiring that they be visible
             to other Habitica users.
  * Feature: Username verification support

Version 4.62.2:

  * API refactor: Move inbox from user object to inboxes collection
  * API fix: Provide API side error when user < level 10 attempts to
             allocate stat points
  * Client fix: Don't show animal ears broken by knockout as purchasable
                with Gems
  * Client fix: Automatically unpin all animal ear rewards when user
                purchases complete set
  * Client fix: Reduce margins in between checklist items on tasks
  * Client fix: Point links in social sharing URLs to Habitica main page
                instead of nonexistent landing pages
  * Client fix: Don't show default Reward value of 1 when user clears
                out value field

Full upstream changelog is at:

https://github.com/HabitRPG/habitica/releases

Signed-off-by: aszlig <aszlig@nix.build>