habitipy: Update and support newer API versions
This currently doesn't yet use the newer API specification yet, because
it will fail with the current Habitica source code, so this is only to
prepare for that.

During my tests while upgrading the Habitica source to the latest
upstream version, I found a few parsing errors with habitipy. Some of
them were errors in Habitica itself, but one was in habitipy not
handling a missing type on apiParams correctly. For the latter I added a
small patch fixing the issue (upstream pull request at [1]).

I also didn't add habiticaSource to the callPackage arguments of
habitipy yet, because it only will become relevant for the upcoming new
upstream version (we have a migration there to test).

[1]: https://github.com/ASMfreaK/habitipy/pull/9

Signed-off-by: aszlig <aszlig@nix.build>