Add patch to fix UUID validation in models
I'm not really sure what's the *exact* reason why this happens (one
webhook test fails because UUID validation doesn't work), but this
commit fixes validation.

Now what the patch does is (v) => theValidator(v) instead of just
theValidator, so what happens here is that just using theValidator
always evaluates to true.

This happens because we have updated a few dependencies to newer
versions and this is also where I'm not sure *which* dependency it is,
but my *guess* is, it might be Babel as I already tried downgrading
validator and mongoose.

While I could spend some more time looking into what the old Babel
versus the old Babel generates and find out the exact reason for this
issue, I'm leaving it like this because I'm time-constrained right now.

Signed-off-by: aszlig <aszlig@nix.build>