Update to version 4.65.3
A much more quiet update compared to the last one, however one thing
stands out, which is a new Google font. I added the font to the
googleFonts derivation, but as reported by @kyren, the derivation is not
really deterministic because Google tends to change these.

The biggest trouble here is that those fonts seem to be autogenerated
based on different browsers - google-fonts-offline basically fakes
different user agent headers to get those fonts.

Other than that, the archival of the inbox migration now is upstream as
well and we can drop the patch. Unfortunately I needed to add another
patch reverting changes in the migration runner, because otherwise the
migrations we had so far won't work.

Upstream API changes:

  * Content: Enable Fall/Halloween themed seasonal avatar skins and hair
             colors for purchase
  * Tests: Added coverage for tasks API routes
  * Chore: Code cleanup related to changes in inbox/private messaging
           data structures
  * Chore: Locale files updated
  * Chore: Bailey news announcements

Upstream client changes:

  * Fix: Don't show class-associated broken items in classless/"None"
         grouping in shop
  * Fix: Better handling and display for parameter validation errors
  * Fix: Correct display for MP gains and losses when allocating
         Intelligence points
  * Fix: Don't show erroneous "you aren't a member of any
         Guilds/Challenges" text if page is not finished loading
  * Fix: Better positioning of @ mention dropdown in text boxes with
         multiple lines of input
  * Fix: Allow "Remove Member" action to appear in context menu when
         accessed from View Party button on Tasks page
  * Fix: Prevent double-click causing extra Gem gift transfers

Signed-off-by: aszlig <aszlig@nix.build>