node-overrides: Fix google fonts hash
As mentioned in an earlier commit, those derivations unfortunately
aren't deterministic. Along the way I also renamed the derivation to the
name 'google-fonts-apidoc' because we already have a derivation with the
same name.

I also removed the name attribute from this derivation and the one in
source.nix, because it was superfluous (it's already passed by the first
argument of the runCommand function).

Signed-off-by: aszlig <aszlig@nix.build>
Reported-by: @kyren