patches: Remove failing test for hasCancelled
The test assumes that hasCancelled will return true whenever there is no
customerId assigned to the plan. However we do have free group plans for
everyone, so hasCancelled will always return false.

Because of this, let's remove the test as we still have a bunch of other
tests (implicitly) asserting the value of hasCancelled.

Signed-off-by: aszlig <aszlig@nix.build>