Update to version 4.67.0
We now no longer need the fix-uuid-validation.patch because it was a
cherry-pick from upstream anyway, so I've dropped that patch on our end.

I'm also glad that the Pusher stuff got removed, which is one less
dependency to take care of and thus we also save a few more hunks in our
patches.

Along the way these direct dependencies have been updated on our side as
well:

  csv-stringify:   3.1.1  -> 4.3.1
  method-override: 2.3.10 -> 3.0.0
  mongoose:        5.3.2  -> 5.3.7

Upstream changes for version 4.67.0:

  * Content: Mystery Items October 2018

Upstream changes for version 4.66.2:

  * Fix: Don't add a Mystery Items notification if new subscriber did
         not receive Mystery Items
  * Fix: Better contrast on reward cost div
  * Fix: Remove extraneous white bar on Group Plans promo page and
         improve text styling
  * Chore: Remove unused Pusher code
  * Chore: Upgrade various packages: mongoose, method-override,
           csv-stringify, node-gcm

Upstream changes for version 4.66.0 and 4.66.1:

  * Fix: Don't reset Gem quantities to 0 when opening Amazon Payments
         modal
  * Content: New veteran pet, Veteran Fox

Signed-off-by: aszlig <aszlig@nix.build>