Fix failures in integration tests for API v4
With the recent refactoring of the user/display name handling, a few new
test cases were added for testing whether slurs are disallowed properly.

However on our side, we don't censor any slurs or banned words, so the
new test cases needed to be fixed in a way to instead check that
whenever a slur is used the user or display name is still allowed.

Signed-off-by: aszlig <aszlig@nix.build>