Fix E2E test for imageproxy
With the recent update, creating a new user will have an additional
modal before the character creation for verification of the user and
display name. Our E2E test for the imageproxy didn't yet account for
this, so by this commit this is now no longer the case (we just need to
skip that modal).

Signed-off-by: aszlig <aszlig@nix.build>