Fix E2E test for imageproxy for real this time
I shouldn't actually try to fix tests when I'm distracted with something
else and the test I was going to fix actually has failed the last time
around. So here is the real fix for the test, which involves one less
step for the character creation, which was the second change that I
missed for the last upstream update.

Signed-off-by: aszlig <aszlig@nix.build>