Update to Node.js 10.x
The upstream project has switched to Node.js 10 in version 4.74.5, so in
order to stay compatible, let's switch to version 10 as well.

Along the way I made it easier to override the node version in one place
only (shabitica.nix) instead of needing to update multiple files in the
future.

I also added a compatibility version of Node.js 10 for NixOS 18.03,
because version 10 isn't included there.

Signed-off-by: aszlig <aszlig@nix.build>