Fix Node.js 10 fallback in restricted-eval mode
While this would work fine on systems without the restrict-eval Nix
option set to true, it doesn't for systems that have it enabled, like
it's the case for Hydra.

So let's coerce the path to nixpkgs to a string, so we hit exactly the
same path as we have in the source argument (which for shabitica.nix is
implicit) and thus don't trigger the restriction.

Fortunately we don't need to explicitly coerce to a path, because
callPackage already imports it if it is *not* a function.

Signed-off-by: aszlig <aszlig@nix.build>