Update to version 4.75.1
A more substantial update when it comes to our patches:

  * Fix: Put Username in account deletion feedback emails and improve
         code
  * Refactor: Change environment variables from nested object format to
              flat keys for better compatibility
  * Chore: Locale files updated
  * Fix: Polish various payment checkout flows, with an emphasis on
         Amazon, for a smoother experience with more feedback and fewer
         page refreshes

So there are two main things affecting us: The move to the flat
structure for the config and some changes in the router, which adds a
new /redirect route which we don't need because it's payment-related.

Another thing which wasn't mentioned in the changelog was the removal of
the leftover code for the Kafka queue, so we no longer need to remove
the code on our side.

Signed-off-by: aszlig <aszlig@nix.build>