Update to version 4.77.5
Relevant upstream updates since version 4.75.5:

  * Content: Winter Wonderland 2019, including Seasonal Shop, new class
    outfits, and the return of the Santa quests
  * Content: New premium Hatching Potion: Icy Snow. Peppermint and
             Starry Night potions available again
    Content: December 2018 Subscriber Mystery Items
  * Fix: Change CSS so close icons show up consistently across site
  * Fix: Adjustments to Penguin and Wolf sprites
  * Feature: When user gifts a subscription, they receive an equivalent
             amount of subscription credit

The were also a bunch of fixes between the releases, but those were only
fixing up the newly introduced features.

One thing that stands out is the last new feature, which we have
reverted on our side. The reason is that it is basically a promotion to
buy Gems and it comes with a nasty nagbar on the top.

We do have free subscriptions for all, so that promo wouldn't make sense
for us, hence I patched it out.

Other than the gift promo issue, the rest of the rebase went smooth and
without conflicts.

Signed-off-by: aszlig <aszlig@nix.build>