Update to version 4.78.0
The changes between 4.77.5 and 4.78.0 are very small and only include a
New Year's Eve event and a small fix in 4.77.6, which removes duplicate
id attributes in the Gems modal. However, the latter doesn't apply to us
because we kill off "buyGemsModal.vue".

Regardless of the rather minor changes I had to rebase the patches,
because the NYE change introduced a new migration script.

Signed-off-by: aszlig <aszlig@nix.build>