Update to version 4.80.6
The changes between version 4.80.0 and 4.80.4 were all about Bailey news
announcements (which we don't have) and the removal of the
"gift-one-get-one" holiday promotion, here is the summary of 4.80.1 to
4.80.4 in regards to removing the promo:

  * Remove gift-one-get-one holiday promotion from subscriptions
  * Remove mentions of gift-one-get-one holiday promotion (banners,
    etc.) from client components
  * Remove missed G1G1 banner from Gem/Subscription purchase modal

Version 4.80.6 is just yet another Bailey news announcement but 4.80.5
contained a very minor but somewhat relevant change:

  * Change SSL check by checking for "http" in x-forwarded-for instead
    of checking for "not 'https'"

So a lot of versions for a not very substantial changes. Of course, our
patch reverting the "gift-one-get-one" promo is dropped with this very
commit.

Signed-off-by: aszlig <aszlig@nix.build>