Update to version 4.81.0
The main content update for 4.81.0 are the new subscriber items, which
only are relevant for us when it comes to the Time Traveller's shop.

Apart from that, the versions from 4.80.7 to 4.80.9 contained the
following fixes:

  * Special pets you don't own show grey pawprint
  * New password can be set with Enter key in password reset flow
  * Do not show "Notification not found" error messages
  * Repair negative quest scrolls on purchase

Note that the last two items here didn't show up in the upstream release
notes, so I added them here accordingly.

In 4.80.7, the following feature was introduced:

  Add the ability to set a secret key which, when passed as a query
  parameter to a request, will proceed without SSL redirection. This
  allows health checks on server architecture to return results other
  than HTTP 302

On our side however, I added a patch to remove this feature, because we
really don't need it and there really is no reason to introduce
complexity for something we don't need.

Signed-off-by: aszlig <aszlig@nix.build>