Update to version 4.82.0
Upstream API changes:

  * Content: New cosmetic armor item, Outlandish Party Robes
  * Feature: (Temporary) Award Outlandish Party Robes to new accounts
  * Feature: (Temporary) Market sells Cake, and Cake randomly drops,
                         instead of usual food items
  * Chore: Bailey news

Upstream client changes:

  * Feature: (Temporary) Display birthday party version of NPC images

Upstream repository changes:

  * Migration: Award party robes and cake to all accounts

Contrary to all the subscriber item migrations, I'm adding this to our
migration list because *all* users are getting the special birthday
items.

The new migrations are returning a Promise, so I've altered and renamed
the revert-migration-runner-promise.patch so that we can run both the
old style migrations that return a value and the new ones returning a
Promise.

For the migration test, I also needed to add another user called
"birthday2019" to our fixture, because only users logged in after
2019-01-15 will be awarded with the birtday gear and the other two test
users are created in 2018.

Signed-off-by: aszlig <aszlig@nix.build>