Get rid of deprecation warning about moment().zone
This is very annoying and it not only shows up during server startup but
also whenever you open the JS console on the client. While this should
ideally be fixed, I disabled the deprecation warning for now, because we
are aware of that and upstream is as well[1].

Unfortunately I had to override this twice, because node2nix currently
simply unpacks the upstream source into the moment and the moment-recur
derivation.

[1]: https://github.com/HabitRPG/habitica/issues/10209

Signed-off-by: aszlig <aszlig@nix.build>