node-overrides: Fix sed expr for gulp-imagemin
The previous sed expression replaced all of the paths with the one from
gifsicle, because we used an address range ending with a match on the
name.

However, sed continues to match each line even if there is no occurence
of the ending address match, which means that in turn *all* blocks are
processed.

In order to fix this, I'm now no longer using an address range but check
for the first line after the match of "new BinWrapper()" and depending
on that match we replace the ".dest" calls later with the right paths.

To illustrate this, I've marked the second line after "new BinWrapper()"
here for the optipng-bin node package:

     module.exports = new BinWrapper()
  ->     .src(url + 'macos/optipng', 'darwin')
         ... other calls to src() ...
         .dest(path.join(__dirname, '../vendor'))

Signed-off-by: aszlig <aszlig@nix.build>