Update to version 4.85.5
The only change relevant for us in version 4.85.5 is the removal of the
Cupid and Rose Quartz potions, but at least it's a bit more relevant
than version 4.85.3 and 4.85.4, where the former only adds a few tweaks
to Stackdriver and the latter allows the "Hall of Heroes" to be
searched and fixes the rendering of the "Apply Sort Options" link.

In our case, the "Hall of Heroes" is hidden from the user, so it's not
relevant, which leavos the "Apply Sort Options" fix and the mentioned
removal of the Cupid and Rose Quartz potions the only relevant changes.

I'd guess that there will be more Stackdriver changes coming up, so I
added it to the list of canaries, ensuring that nothing about that gets
sneaked in, should I one day stop following the upstream changes.

Signed-off-by: aszlig <aszlig@nix.build>