Minify Google fonts and strip comments
Unfortunately we got a hash mismatch for the Google fonts again, so I
decided to check what has actually changed and it turns out that the
changes were just a few comments and no actual data.

So instead of just updating the hashes, I opted for stripping the
comments and minimize the CSS so that we hopefully have a predictable
output.

This doesn't make the fonts entirely deterministic though, so for
example if one of the actual font file changes we'll still end up with a
hash mismatch.

Hopefully by the time this happens we already have a better solution,
where we can reconstruct it from the "google-fonts" package.

Thanks to @kyren for reporting this. Luckily there are some people out
there who garbage collect more often than I do, so they'll notice X-)

Signed-off-by: aszlig <aszlig@nix.build>
Closes: https://github.com/headcounter/shabitica/pull/3