Add a few headers employing a more strict policy
So far we containerised our backend and also made sure that we remove
all requests to external sites by patching the source.

However, to be absolutely sure this will never happen, let's actually
set a content security policy header to the clients, so that they won't
load anything from external sites, in case we forget about something we
need to patch.

Ideally, we'd also have a backend node which will report this back to
the user (rather than just being reported in the JS console), so that we
can actually fix this. However not actually loading such content is
already an improvement.

In addition, I added a referrer policy, so that browsers don't send a
"Referer" header once users will visit external sites. We had this
already in a few <a/> tags, but specifying it in headers makes sure this
is consistent.

The last header (X-Content-Type-Options) makes sure we don't load styles
and scripts with the wrong content type. With the CSP in place and given
that our own backend shouldn't deliver something malicious, this
probably won't happen, but let's better be safe than sorry.

Signed-off-by: aszlig <aszlig@nix.build>