Further restrict the content security policy
We also don't (usually) want Shabitica to be embedded via eg. an iframe
from another page, which would open a whole lot of other ways to trick
users.

However, in some cases it's legitimate to allow this for some sites, so
I added a new allowEmbedFrom option to allow users to configure this.

Also, Shabitica shouldn't use plugins, so I also explicitly disallowed
the use of <object/>.

Signed-off-by: aszlig <aszlig@nix.build>