Move helper tools into dedicated subdirectory
The top-level directory already gets messy enough, so let's get a few
things out of the way.

Furthermore, this moves the descriptions about the tools into the
README.rst inside the new "tools" directory, because using these tools
is only related to maintenance and development of this fork and thus
aren't really relevant in the top-level README.rst.

Signed-off-by: aszlig <aszlig@nix.build>