Update to version 4.91.2
Upstream API changes:

  * Feature: Add the option to use pagination for private messages
  * Refactor: Overhaul of user model structure and validation for user
              inventory to boost performance

Upstream client changes:

  * Fix: Add a min-width to private messages to help properly line-break
         long URLs
  * Fix: Add various failsafes to ensure that pinned items don't break
         the Rewards column

A few more conflicts showed up when rebasing the patches compared to the
last rebase, however those were mostly because code was moved around
during the overhoul of the user model structure and the update of a
bunch of dependencies.

Due to the last point I also updated the node dependencies on our side
and made sure that mongoose is at 5.4.19, because the dependency got
downgraded upstream from 5.4.20 to 5.4.19.

I also had to fix our inboxOutsideUser migration test, because querying
the inbox messages now requires a page argument.

Signed-off-by: aszlig <aszlig@nix.build>