Fix nginx configuration errors reported by gixy
The first issue reported is a false-positive reported on the /apidoc
URL. Usually if the location is /apidoc and the alias points to
something like /foo/bar/, an attacker can retrieve a file /foo/other by
requesting the URL https://example.com/apidoc../other.

In our case however, this doesn't work, because the directory given to
the alias directive doesn't end with a slash, so requesting
https://example.com/apidoc../other from the example above will try to
retrieve /foo/bar../other.

Of course, this is still not very nice, because if we have a store path
with the same hash but a different name (which is only the case if we
have a SHA256 collision), the attacker could retrieve the contents of
the other path.

To illustrate what I mean, let's say there is a file called
/foo/barsecret, the attacker could get it via
https://example.com/apidocsecret.

So to shut up gixy and also prevent us from accidentally appending a
slash to the apidoc store path, I've fixed this accordingly.

Another problem reported by gixy was that the $http_host header is prone
to spoofing attacks. However this isn't a problem with Shabitica anyway,
because we don't have any instances in the code where we check the host.

I also fixed the latter however, because we can't guarantee that the
host header is going to be irrelevant indefinitely.

Signed-off-by: aszlig <aszlig@nix.build>