README: Add reference to the NixOS 19.03 channel
NixOS 19.03 has been released since quite a while, so while the Hydra
jobset already exist, we might want to actually refer to it from the
project's description.

I also removed the reference to NixOS 18.03. It still has a Hydra
channel and a corresponding jobset, so it will continue to be supported
for a while, but we might start breaking things sooner or later, hence
we don't mention it anymore in the README.

Signed-off-by: aszlig <aszlig@nix.build>