migrations: Fix test of inboxOutsideUser migration
Since version 4.93.3, the inbox API has changed and for retrieving the
messages, there now is a new "conversation" parameter which expects the
UUID of the other user of a series of messages the current user is
participating.

The parameter is required, so we need to supply it in our tests as well.

Signed-off-by: aszlig <aszlig@nix.build>