Update to version 4.94.1
Version 4.94.0 just adds new Enchanted Armoire items and backgrounds for
May 2019, so it's solely a content update.

Fixes and new features are introduced in 4.94.1, which are:

  API:

    * Feature: Add specific conversation option to private messages GET
    * Content: Re-enable Feathered Friends Bundle
    * Fix: Correct issue where owned challenges list would come up empty
    * Chore: Bailey news re quest bundle

  Client:

    * Fix: Correct task column width/centering in certain viewport x
           language combinations
    * Fix: Don't show seasonal time-limited warnings on items you
           already own
    * Fix: Reload page after activating Orb of Rebirth, to catch the
           many items needing
    * Fix: Correct wrapping in tags with long text
    * Fix: Group tags from challenges, groups, and user in sections when
           editing task
    * Fix: Add infinite-scroll loading support to Challenges
    * Fix: Improve skill usage performance by limiting database fields
           queried
    * Fix: Improve color gradients on Dojo background

Rebasing the patches was a bit more involved, because what was omitted
from the upstream changelog is a bit of a refactor of cancellation of
group plans. I updated our patches accordingly and also added the new
modals to the list of files to remove.

In addition to that, I also added billing.vue to the kill list, because
after inspecting the source I found out that it was unreferenced
(especially after removing the cancellation modals).

Unfortunately, the tests do not pass (even upstream) by just using the
release version, so I essentially cherry-picked a commit that fixes the
date range of the new quest bundle which we should be able to drop again
with the next upstream version.

Signed-off-by: aszlig <aszlig@nix.build>