build-vm: Fix race condition while waiting for SSH
So far we only checked whether the TCP port of the SSH server was
accepting connections. However there is a small time frame where SSH
just closes the connection if it's not ready yet.

If we happen to get in between that time frame, the VM immediately shuts
down again, so let's actually wait until we get some SSHd server
software version and only continue if that's the case.

Signed-off-by: aszlig <aszlig@nix.build>