Add ability to generate invite URLs
On the server side, we don't necessarily have an MTA running, so
inviting others via email is not an option. This is getting even more
crucial because by default we are in invite-only mode.

In addition, people not always have the email address of another person
but rather just a contact via some other means (eg. instant messaging),
so just being able to paste an URL is way more convenient.

The implementation adds a new API endpoint (/groups/ID/generate-invite),
which just returns a new invitation URL and also isn't subject to the
invitation limit, because generating these URLs is a way cheaper
operation than sending out one or more emails.

This is addressing a comment[1] from @Runear:

  I realise you can only add people to a server by invite, but that
  invite function for me isn't working. At least I don't think it is. I
  tried to send an invite from within the app and nothing comes through.

While I don't know his configuration, my guess is that a configuration
for an MTA was missing, so having a way to directly get invite URLs we
now no longer necessarily need to have an MTA.

[1]: https://github.com/headcounter/shabitica/issues/5#issuecomment-487263060

Signed-off-by: aszlig <aszlig@nix.build>
Reported-by: https://github.com/Runear